Machine_WEB/api/requirements.txt
mysql-connector-python==8.0.33
python-dotenv==1.0.0
pandas==2.1.4  # Version spécifique pour compatibilité
numpy==1.26.4  # Version 1.x stable
pysnmp==4.4.12
requests==2.31.0
beautifulsoup4==4.12.2
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
colorama==0.4.6
fastapi==0.115.12
h11==0.14.0
idna==3.10
pydantic==2.11.1
pydantic_core==2.33.0
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.0
uvicorn==0.34.0
ultralytics==8.3.101
opencv-python-headless==4.8.0.76
mysql-connector-python==8.0.33  # Driver officiel (performances optimales)
PyMySQL==1.1.0                  # Alternative pure Python (meilleure compatibilité cross-platform)
httpx==0.28.1
ffmpeg-python==0.2.0
aiohttp==3.11.16
yt-dlp==2025.3.31
Unidecode==1.3.8
websockets==15.0.1